imsp/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
mysqlclient==2.1.0
python-decouple==3.6
sqlparse==0.4.2
